+++ Runtime/ObjPool/IDisposableDebugTarget.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace OcUtility
+{
+    public interface IDisposableDebugTarget
+    {
+        Type GetGenericType();
+        void DebugType();
+        void DebugLog();
+        void Dispose();
+    }
+}